fix: formatting

+++ test/unit/concrete/yieldbox/YieldBox.t.sol
@@ -45,18 +45,9 @@ contract YieldBoxUnitConcreteTest is BaseTest {
         super.setUp();
 
         // Create strategies
-        daiStrategy = new ERC20WithoutStrategy(
-            IYieldBox(address(yieldBox)),
-            IERC20(address(dai))
-        );
-        wrappedNativeStrategy = new ERC20WithoutStrategy(
-            IYieldBox(address(yieldBox)),
-            IERC20(address(wrappedNative))
-        );
-        usdtStrategy = new ERC20WithoutStrategy(
-            IYieldBox(address(yieldBox)),
-            IERC20(address(usdt))
-        );
+        daiStrategy = new ERC20WithoutStrategy(IYieldBox(address(yieldBox)), IERC20(address(dai)));
+        wrappedNativeStrategy = new ERC20WithoutStrategy(IYieldBox(address(yieldBox)), IERC20(address(wrappedNative)));
+        usdtStrategy = new ERC20WithoutStrategy(IYieldBox(address(yieldBox)), IERC20(address(usdt)));
 
         // Register assets in YieldBox
 
@@ -70,19 +61,11 @@ contract YieldBoxUnitConcreteTest is BaseTest {
 
         // Register Wrapped native
         WRAPPED_NATIVE_ASSET_ID = yieldBox.registerAsset(
-            TokenType.ERC20,
-            address(wrappedNative),
-            IStrategy(address(wrappedNativeStrategy)),
-            0
+            TokenType.ERC20, address(wrappedNative), IStrategy(address(wrappedNativeStrategy)), 0
         );
 
         // Register USDT
-        USDT_ASSET_ID = yieldBox.registerAsset(
-            TokenType.ERC20,
-            address(usdt),
-            IStrategy(address(usdtStrategy)),
-            0
-        );
+        USDT_ASSET_ID = yieldBox.registerAsset(TokenType.ERC20, address(usdt), IStrategy(address(usdtStrategy)), 0);
 
         // Prank impartial user
         _resetPrank({msgSender: users.alice});
@@ -93,20 +76,8 @@ contract YieldBoxUnitConcreteTest is BaseTest {
     /////////////////////////////////////////////////////////////////////
 
     /// @notice Modifier utilized to perform a single deposit into YieldBox.
-    modifier whenDeposited(
-        uint256 _assetId,
-        address _from,
-        address _to,
-        uint256 _amount,
-        uint256 _share
-    ) {
-        _whenDeposited({
-            assetId: _assetId,
-            from: _from,
-            to: _to,
-            amount: _amount,
-            share: _share
-        });
+    modifier whenDeposited(uint256 _assetId, address _from, address _to, uint256 _amount, uint256 _share) {
+        _whenDeposited({assetId: _assetId, from: _from, to: _to, amount: _amount, share: _share});
         _;
     }
 
@@ -115,36 +86,17 @@ contract YieldBoxUnitConcreteTest is BaseTest {
     ///     - DAI
     ///     - WRAPPED NATIVE
     ///     - USDT
-    modifier whenDepositedAll(
-        address _from,
-        address _to,
-        uint256 _amount,
-        uint256 _share
-    ) {
+    modifier whenDepositedAll(address _from, address _to, uint256 _amount, uint256 _share) {
         for (uint256 assetID = 1; assetID <= 3; assetID++) {
-            _whenDeposited({
-                assetId: assetID,
-                from: _from,
-                to: _to,
-                amount: _amount,
-                share: _share
-            });
+            _whenDeposited({assetId: assetID, from: _from, to: _to, amount: _amount, share: _share});
         }
 
         _;
     }
 
     /// @notice Modifier utilized to perform multiple deposits into YieldBox.
-    modifier simulateYieldBoxDeposits(
-        uint256 _assetId,
-        uint256 _amount,
-        uint256 _share
-    ) {
-        _simulateYieldBoxDeposits({
-            assetId: _assetId,
-            amount: _amount,
-            share: _share
-        });
+    modifier simulateYieldBoxDeposits(uint256 _assetId, uint256 _amount, uint256 _share) {
+        _simulateYieldBoxDeposits({assetId: _assetId, amount: _amount, share: _share});
         // Reset prank to default impartial user
         _resetPrank(users.alice);
         _;
@@ -154,13 +106,7 @@ contract YieldBoxUnitConcreteTest is BaseTest {
     //                       INTERNAL HELPERS                          //
     /////////////////////////////////////////////////////////////////////
     /// @notice Function utilized to perform a single deposit into YieldBox.
-    function _whenDeposited(
-        uint256 assetId,
-        address from,
-        address to,
-        uint256 amount,
-        uint256 share
-    ) internal {
+    function _whenDeposited(uint256 assetId, address from, address to, uint256 amount, uint256 share) internal {
         // Approve yieldBox to transfer `from` assets
         _approveViaPearlmit({
             from: from,
@@ -170,24 +116,14 @@ contract YieldBoxUnitConcreteTest is BaseTest {
         });
 
         // Deposit assets in YieldBox
-        yieldBox.depositAsset({
-            assetId: assetId,
-            from: from,
-            to: to,
-            amount: amount,
-            share: share
-        });
+        yieldBox.depositAsset({assetId: assetId, from: from, to: to, amount: amount, share: share});
     }
 
     /// @notice Function utilized to perform several deposits with different users into YieldBox.
     /// @dev Note that `alice` does not deposit.
-    function _simulateYieldBoxDeposits(
-        uint256 assetId,
-        uint256 amount,
-        uint256 share
-    ) internal {
+    function _simulateYieldBoxDeposits(uint256 assetId, uint256 amount, uint256 share) internal {
         // Skip alice address
-        for (uint i = 1; i < userAddresses.length; i++) {
+        for (uint256 i = 1; i < userAddresses.length; i++) {
             _whenDeposited({
                 assetId: assetId,
                 from: userAddresses[i],
@@ -216,31 +152,11 @@ contract YieldBoxUnitConcreteTest is BaseTest {
     ) internal view returns (uint8 v, bytes32 r, bytes32 s) {
         // Build struct hash for YieldBox permit
         bytes32 structHash = isForAssetId
-            ? keccak256(
-                abi.encode(
-                    isPermit ? PERMIT_TYPEHASH : REVOKE_TYPEHASH,
-                    owner,
-                    spender,
-                    assetId,
-                    nonce,
-                    deadline
-                )
-            )
-            : keccak256(
-                abi.encode(
-                    isPermit ? PERMIT_ALL_TYPEHASH : REVOKE_ALL_TYPEHASH,
-                    owner,
-                    spender,
-                    nonce,
-                    deadline
-                )
-            );
+            ? keccak256(abi.encode(isPermit ? PERMIT_TYPEHASH : REVOKE_TYPEHASH, owner, spender, assetId, nonce, deadline))
+            : keccak256(abi.encode(isPermit ? PERMIT_ALL_TYPEHASH : REVOKE_ALL_TYPEHASH, owner, spender, nonce, deadline));
 
         // Convert to typed data hash
-        bytes32 hash = ECDSA.toTypedDataHash(
-            yieldBox.DOMAIN_SEPARATOR(),
-            structHash
-        );
+        bytes32 hash = ECDSA.toTypedDataHash(yieldBox.DOMAIN_SEPARATOR(), structHash);
 
         // Sign the typed data hash
         (v, r, s) = vm.sign(privateKey, hash);
